fix : Minor changes
